moved: goal tile components into their components dir
