+++ src/components/GoalsComponents/MyGoal/components/GoalDropdown.tsx
@@ -9,21 +9,16 @@ interface GoalDropdownProps {
 const GoalDropdown: React.FC<GoalDropdownProps> = ({ goal }) => {
   const { sublist, goalColor, timeBudget, newUpdates } = goal;
   const hasSubGoals = sublist.length > 0;
+
   const outerBackground = `radial-gradient(50% 50% at 50% 50%, ${goalColor}33 89.585%, ${
     timeBudget?.perDay != null ? "transparent" : goalColor
   } 100%)`;
-  const outerBorderStyle = timeBudget?.perDay == null ? `1px solid ${goalColor}` : `2px dashed ${goalColor}`;
 
   const innerBorderColor = hasSubGoals ? goalColor : "transparent";
+  const outerBorderStyle = timeBudget?.perDay == null ? `1px solid ${goalColor}` : `2px dashed ${goalColor}`;
 
   return (
-    <div
-      style={{
-        display: "flex",
-        flexDirection: "column",
-        gap: 3,
-      }}
-    >
+    <div className="d-flex f-col gap-4">
       <div
         className="goal-dropdown goal-dd-outer"
         style={{
